html/browsers/windows/nested-browsing-contexts/frameElement-02.html

<!DOCTYPE html>
<meta charset="utf-8"/>
<title>HTML Test: window.frameElement (window created by the frame element)</title>
<link rel="author" title="Intel" href="http://www.intel.com/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var t1 = async_test("The window's frameElement attribute must return its container element if it is a nested browsing context")
function on_load() {
  t1.step(function () {
    assert_equals(document.getElementById("fr1").contentWindow[0].frameElement,
                  document.getElementById("fr1").contentDocument.getElementById("f1"),
                 "The frameElement attribute should be the frame element in 'test.html'.");
  });
  t1.done();
}
</script>
<iframe id="fr1" src="test.html" onload="on_load()" style="display:none"></iframe>
